tests(user): improve user registration tests
Added new tests for successful registration, redirection and
user data saving. Improved coverage for user registration
scenario.
I thought about the possibility of testing when user registered
and was redirected to the correct page with a valid URL.
This is important to avoid erros 404 and 500 when registering in
the system.